// src/services/openapi/auth-service/api.ts
export interface Division {
    /**
     * 
     * @type {string}
     * @memberof Division
     */
    createdDate?: string;
    /**
     * 
     * @type {string}
     * @memberof Division
     */
    lastModifiedDate?: string;
    /**
     * 
     * @type {number}
     * @memberof Division
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Division
     */
    divisionCode?: string;
    /**
     * 
     * @type {Category}
     * @memberof Division
     */
    category?: Category;
    /**
     * 
     * @type {Directorate}
     * @memberof Division
     */
    directorate?: Directorate;
    /**
     * 
     * @type {Group}
     * @memberof Division
     */
    group?: Group;
    /**
     * 
     * @type {string}
     * @memberof Division
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof Division
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof Division
     */
    segmentFinance?: string;
}
// ...
export interface GroupMenuApps {
    /**
     * 
     * @type {string}
     * @memberof GroupMenuApps
     */
    createdDate?: string;
    /**
     * 
     * @type {string}
     * @memberof GroupMenuApps
     */
    lastModifiedDate?: string;
    /**
     * 
     * @type {number}
     * @memberof GroupMenuApps
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof GroupMenuApps
     */
    groupMenuCode?: string;
    /**
     * 
     * @type {string}
     * @memberof GroupMenuApps
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof GroupMenuApps
     */
    description?: string;
}
/**
 * 
 * @export
 * @interface LoginRequestDto
 */
export interface LoginRequestDto {
    /**
     * 
     * @type {string}
     * @memberof LoginRequestDto
     */
    username?: string;
    /**
     * 
     * @type {string}
     * @memberof LoginRequestDto
     */
    password?: string;
}
/**
 * 
 * @export
 * @interface LoginResponseDto
 */
export interface LoginResponseDto {
    /**
     * 
     * @type {string}
     * @memberof LoginResponseDto
     */
    token?: string;
    /**
     * 
     * @type {boolean}
     * @memberof LoginResponseDto
     */
    needPassword?: boolean;
    /**
     * 
     * @type {number}
     * @memberof LoginResponseDto
     */
    totalAttemptsPasswordFailed?: number;
    /**
     * 
     * @type {boolean}
     * @memberof LoginResponseDto
     */
    otp?: boolean;
    /**
     * 
     * @type {string}
     * @memberof LoginResponseDto
     */
    redirectSSO?: string;
    /**
     * 
     * @type {UserDetailResponse}
     * @memberof LoginResponseDto
     */
    user?: UserDetailResponse;
    /**
     * 
     * @type {Array<RoleDto>}
     * @memberof LoginResponseDto
     */
    listRole?: Array<RoleDto>;
    /**
     * 
     * @type {Array<UserGroupDto>}
     * @memberof LoginResponseDto
     */
    listUserGroup?: Array<UserGroupDto>;
    /**
     * 
     * @type {boolean}
     * @memberof LoginResponseDto
     */
    needChangePassword?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof LoginResponseDto
     */
    ldap?: boolean;
    /**
     * 
     * @type {{ [key: string]: object; }}
     * @memberof LoginResponseDto
     */
    additionalData?: { [key: string]: object; };
}
// ...
export interface OtpResponseDto {
    /**
     * 
     * @type {string}
     * @memberof OtpResponseDto
     */
    token?: string;
    /**
     * 
     * @type {boolean}
     * @memberof OtpResponseDto
     */
    success?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof OtpResponseDto
     */
    expired?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof OtpResponseDto
     */
    locked?: boolean;
    /**
     * 
     * @type {string}
     * @memberof OtpResponseDto
     */
    lockedDate?: string;
    /**
     * 
     * @type {string}
     * @memberof OtpResponseDto
     */
    startDateRequestOtp?: string;
    /**
     * 
     * @type {number}
     * @memberof OtpResponseDto
     */
    remaining?: number;
    /**
     * 
     * @type {number}
     * @memberof OtpResponseDto
     */
    totalAttempts?: number;
    /**
     * 
     * @type {number}
     * @memberof OtpResponseDto
     */
    failedAttempts?: number;
    /**
     * 
     * @type {number}
     * @memberof OtpResponseDto
     */
    maxAttempts?: number;
    /**
     * 
     * @type {UserDetailResponse}
     * @memberof OtpResponseDto
     */
    user?: UserDetailResponse;
    /**
     * 
     * @type {Array<RoleDto>}
     * @memberof OtpResponseDto
     */
    listRole?: Array<RoleDto>;
    /**
     * 
     * @type {Array<UserGroupDto>}
     * @memberof OtpResponseDto
     */
    listUserGroup?: Array<UserGroupDto>;
    /**
     * 
     * @type {boolean}
     * @memberof OtpResponseDto
     */
    needChangePassword?: boolean;
    /**
     * 
     * @type {{ [key: string]: object; }}
     * @memberof OtpResponseDto
     */
    additionalData?: { [key: string]: object; };
}
/**
 * 
 * @export
 * @interface Position
 */
export interface Position {
    /**
     * 
     * @type {string}
     * @memberof Position
     */
    createdDate?: string;
    /**
     * 
     * @type {string}
     * @memberof Position
     */
    lastModifiedDate?: string;
    /**
     * 
     * @type {number}
     * @memberof Position
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Position
     */
    positionCode?: string;
    /**
     * 
     * @type {string}
     * @memberof Position
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof Position
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof Position
     */
    division?: string;
    /**
     * 
     * @type {string}
     * @memberof Position
     */
    type?: string;
}
/**
 * 
 * @export
 * @interface RoleDto
 */
export interface RoleDto {
    /**
     * 
     * @type {number}
     * @memberof RoleDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof RoleDto
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof RoleDto
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof RoleDto
     */
    roleCode?: string;
}
/**
 * 
 * @export
 * @interface RoleRefactor
 */
export interface RoleRefactor {
    /**
     * 
     * @type {string}
     * @memberof RoleRefactor
     */
    createdDate?: string;
    /**
     * 
     * @type {string}
     * @memberof RoleRefactor
     */
    lastModifiedDate?: string;
    /**
     * 
     * @type {number}
     * @memberof RoleRefactor
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof RoleRefactor
     */
    roleCode?: string;
    /**
     * 
     * @type {string}
     * @memberof RoleRefactor
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof RoleRefactor
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof RoleRefactor
     */
    group?: string;
    /**
     * 
     * @type {string}
     * @memberof RoleRefactor
     */
    superiorRole?: string;
}
/**
 * 
 * @export
 * @interface UserDetailResponse
 */
export interface UserDetailResponse {
    /**
     * 
     * @type {string}
     * @memberof UserDetailResponse
     */
    fullName?: string;
    /**
     * 
     * @type {string}
     * @memberof UserDetailResponse
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof UserDetailResponse
     */
    nik?: string;
    /**
     * 
     * @type {string}
     * @memberof UserDetailResponse
     */
    status?: UserDetailResponseStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof UserDetailResponse
     */
    statusLabel?: string;
    /**
     * 
     * @type {string}
     * @memberof UserDetailResponse
     */
    privyId?: string;
    /**
     * 
     * @type {string}
     * @memberof UserDetailResponse
     */
    skill?: string;
    /**
     * 
     * @type {string}
     * @memberof UserDetailResponse
     */
    lastLoginDate?: string;
    /**
     * 
     * @type {RoleRefactor}
     * @memberof UserDetailResponse
     */
    roleRefactor?: RoleRefactor;
    /**
     * 
     * @type {Array<Division>}
     * @memberof UserDetailResponse
     */
    division?: Array<Division>;
    /**
     * 
     * @type {Array<Position>}
     * @memberof UserDetailResponse
     */
    position?: Array<Position>;
    /**
     * 
     * @type {UserDetailSubData}
     * @memberof UserDetailResponse
     */
    superior?: UserDetailSubData;
    /**
     * 
     * @type {Array<GroupMenuApps>}
     * @memberof UserDetailResponse
     */
    groupMenuApps?: Array<GroupMenuApps>;
    /**
     * 
     * @type {string}
     * @memberof UserDetailResponse
     */
    userId?: string;
    /**
     * 
     * @type {string}
     * @memberof UserDetailResponse
     */
    userFrom?: UserDetailResponseUserFromEnum;
    /**
     * 
     * @type {string}
     * @memberof UserDetailResponse
     */
    bucketProcessId?: string;
}

/**
    * @export
    * @enum {string}
    */
export enum UserDetailResponseStatusEnum {
    ACTIVE = 'ACTIVE',
    INACTIVE = 'IN_ACTIVE',
    BLOCKEDOTPREQUEST = 'BLOCKED_OTP_REQUEST',
    SUSPEND = 'SUSPEND'
}
/**
    * @export
    * @enum {string}
    */
export enum UserDetailResponseUserFromEnum {
    NEWLOS = 'NEW_LOS',
    LDAP = 'LDAP',
    NONLDAP = 'NON_LDAP',
    INTERNAL = 'INTERNAL',
    EXTERNAL = 'EXTERNAL'
}

/**
 * 
 * @export
 * @interface UserDetailSubData
 */
export interface UserDetailSubData {
    /**
     * 
     * @type {string}
     * @memberof UserDetailSubData
     */
    fullName?: string;
    /**
     * 
     * @type {string}
     * @memberof UserDetailSubData
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof UserDetailSubData
     */
    nik?: string;
    /**
     * 
     * @type {string}
     * @memberof UserDetailSubData
     */
    userId?: string;
}
// ...
export interface UserGroupDto {
    /**
     * 
     * @type {number}
     * @memberof UserGroupDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof UserGroupDto
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof UserGroupDto
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof UserGroupDto
     */
    divisionCode?: string;
}